#3 Login completed succesfully with some bugs that need to be fixed then
