+++ src/components/SignInForm/styles.css
@@ -3,6 +3,6 @@ form {
 }
 
 ion-footer {
-    position: absolute;
     bottom: 0;
+    position: sticky;
 }
