+++ src/components/SignUpForm/index.tsx
@@ -9,66 +9,16 @@ import {
   IonInput,
   IonButton,
 } from "@ionic/react";
-import { useForm } from "react-hook-form";
-import { yupResolver } from "@hookform/resolvers/yup";
-import * as yup from "yup";
 
-export enum userTypeEnum {
-  club = "Club",
-  jugador = "Jugador",
-  academia = "Académia",
-  tecnico = "Técnico",
-}
-
-export interface IForm {
-  userType: userTypeEnum;
-  name: string;
-  phone: number;
-  email: string;
-  password: string;
-}
-
-const ERROR_MESSAGES = {
-  required: "Este campo es requerido",
-  positive: "Debe ser un número positivo",
-  email: "Email no válido",
-  number: "Debe especificar un número",
-};
-
-const schema = yup.object().shape({
-  userType: yup.string().required(ERROR_MESSAGES.required),
-  name: yup.string().required(ERROR_MESSAGES.required),
-  phone: yup
-    .number()
-    .typeError(ERROR_MESSAGES.number)
-    .positive(ERROR_MESSAGES.positive)
-    .required(ERROR_MESSAGES.required),
-  email: yup
-    .string()
-    .required(ERROR_MESSAGES.required)
-    .email(ERROR_MESSAGES.email),
-  password: yup.string().required(ERROR_MESSAGES.required),
-});
-
-export const SignUpForm = ({ onSubmit }: any) => {
-  const initialValues = {
-    name: "",
-    email: "",
-    password: "",
-  };
-
-  const {
-    register,
-    handleSubmit,
-    clearErrors,
-    formState: { errors },
-  } = useForm<IForm>({
-    defaultValues: initialValues,
-    resolver: yupResolver(schema),
-  });
+export const SignUpForm = ({
+  register,
+  handleSubmit,
+  clearErrors,
+  errors,
+}: any) => {
 
   return (
-      <form onSubmit={handleSubmit(onSubmit)}>
+      <form onSubmit={handleSubmit}>
         <IonRow>
           <IonCol size="12">
             <IonItem>
