+++ src/pages/PlayerDashboard/PlayerDashboard.tsx
@@ -1,5 +1,4 @@
 import {
-  IonBackButton,
   IonButtons,
   IonCol,
   IonContent,
@@ -26,7 +25,7 @@ const PlayerDashboard: React.FC = () => (
         </IonRow>
       </IonToolbar>
     </IonHeader>
-    <IonContent fullscreen className="ion-padding"></IonContent>
+    <IonContent fullscreen className="ion-padding">hello word I'm sad cause a can't find the problem :'c</IonContent>
   </IonPage>
 );
