+++ src/pages/SignIn/styles.css
@@ -23,3 +23,17 @@
   font-size: 26px;
   font-weight: bold;
 }
+
+.signin-button {
+  border-radius: 10px;
+  background-color: var(--ion-color-primary);
+  color: #FFFFFF;
+  padding-right: 1.1em;
+  padding-left: 1.1em;
+  letter-spacing: 0.84px;
+  height: 36px;
+  font-size: 14px;
+  font-weight: 500;
+  font-family: var(--ion-font-family, inherit);
+  box-shadow: 0 3px 1px -2px rgba(0, 0, 0, 0.2), 0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 1px 5px 0 rgba(0, 0, 0, 0.12);
+}
